Fix: fire_eye's damage zone
Damage 0 is no longer displayed
An array that remembers who is inside and while this mob is inside it sends a signal that the enemy needs to be damaged.
As soon as this body leaves the eye area, it is removed from the array and is no longer to be damaged.